+++ src/Tool/Wrapper/AbstractWrapper.php
@@ -89,6 +89,11 @@ abstract class AbstractWrapper implements WrapperInterface
 
     public function populate(array $data)
     {
+        @trigger_error(sprintf(
+            'Using "%s()" method is deprecated since gedmo/doctrine-extensions 3.x and will be removed in version 4.0.',
+            __METHOD__
+        ), E_USER_DEPRECATED);
+
         foreach ($data as $field => $value) {
             $this->setPropertyValue($field, $value);
         }